requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
babelfish==0.6.1
blinker==1.9.0
cachetools==7.0.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cinemagoer==2023.5.1
click==8.3.1
colorama==0.4.6
dataclass-factory==2.16
decorator==5.2.1
docutils==0.22.4
easyocr==1.7.2
faiss-cpu==1.13.2
ffmpeg-python==0.2.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
guessit==3.8.0
huggingface-hub==0.36.0
id==1.6.1
idna==3.11
ImageIO==2.37.3
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==25.7.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mutagen==1.47.0
narwhals==2.18.0
networkx==3.6.1
nh3==0.3.3
ninja==1.13.0
numba==0.63.1
numpy==2.3.5
opencv-python-headless==4.13.0.92
packaging==25.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.5.1
pooch==1.8.2
progress==1.6.1
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pyclipper==1.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
readme_renderer==44.0
rebulk==3.2.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
shazamio==0.8.1
shazamio_core==1.1.2
six==1.17.0
smmap==5.0.3
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.45
static_ffmpeg==3.0
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2026.3.3
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.1
transformers==4.57.5
twine==6.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
yarl==1.22.0
